Move some files into components
