+++ src/renderer/actions/ActionPlaySound.tsx
@@ -5,7 +5,7 @@ import {produce} from "immer";
 import { useSelector } from "react-redux";
 import { Icon } from "semantic-ui-react";
 
-import { InlineDropdown } from "@/containers/Misc/InlineInputs";
+import { InlineDropdown } from "@/components/InlineInputs";
 import { AddSoundButton } from "@/containers/Settings/SoundSettings";
 import { soundPlayer } from "@/lib/sounds";
 import { iRootState } from "@/store";
