+++ src/renderer/actions/ActionWriteFile.tsx
@@ -6,7 +6,7 @@ import { ActionTypeGenerator, Context } from "@vinceau/event-actions";
 import { produce } from "immer";
 import { Form, Icon, TextArea } from "semantic-ui-react";
 
-import { InlineDropdown } from "@/containers/Misc/InlineInputs";
+import { InlineDropdown } from "@/components/InlineInputs";
 import { FileInput } from "@/containers/Misc/Misc";
 import { notify as sendNotification } from "@/lib/utils";
 import { writeFile } from "common/utils";
