+++ src/renderer/components/CharacterIcon.tsx
