+++ src/renderer/components/InlineInputs.tsx
