+++ src/renderer/containers/Settings/ComboFinder.tsx
@@ -7,7 +7,7 @@ import { fileProcessor, FindComboOption, ProcessResult } from "@/lib/fileProcess
 import { loadFileInDolphin, notify, openComboInDolphin } from "@/lib/utils";
 import { Dispatch, iRootState } from "@/store";
 import { secondsToString } from "common/utils";
-import { InlineDropdown } from "../Misc/InlineInputs";
+import { InlineDropdown } from "../../components/InlineInputs";
 import { FileInput } from "../Misc/Misc";
 import { ProcessSection } from "../Misc/ProcessSection";
 import { RenameFiles } from "./RenameFiles";
