+++ src/renderer/containers/Settings/Settings.tsx
@@ -15,7 +15,8 @@ import styled, { css } from "styled-components";
 import { device } from "@/styles/device";
 import OBSLogo from "@/styles/images/obs.svg";
 import SlippiLogo from "@/styles/images/slippi-logo.svg";
-import { CustomIcon, Labelled } from "../Misc/Misc";
+import { CustomIcon } from "@/components/CustomIcon";
+import { Labelled } from "../Misc/Misc";
 import { ComboFinder } from "./ComboFinder";
 import { FilterOptions } from "./FilterOptions";
 import { OBSSettings } from "./OBSSettings";
